Update googleSheets.ts
The purpose is to add the 6 fields that exist in the csv export file (csv.ts), but are missing from this Google Sheets export. 
Please note that I used to be a programmer, but in C++, not Java or anything after it.  
I have no idea how to test the changes that I propose ... , so this is NOT TESTED.