Fix: Ensure consistent dates for G-Sheets export (#480)
This is a fix for https://github.com/brafdlog/caspion/issues/476
("google spreadsheet dates inconsistent cell format")

The fix is to use the format of `YYYY-MM-DD` which has no ambiguity and
Google Sheets will automatically detect it as a date field.

I did not see any unit-testing in this project so added no tests. If I
am missing the tests, I will add.