Fix: Ubuntu GitHub test action (#643)
The `ubuntu-latest` tag was changed to point to Ubuntu 24.04 and has
caused regressions. See
https://github.com/microsoft/playwright/issues/34251 and
https://github.com/electron/electron/issues/42510 for more details.
TL;DR: 
> The solution is to lift the restrictions that Ubuntu 24.04 implements
in the AppImages.
> ```
> sudo sysctl -w kernel.apparmor_restrict_unprivileged_userns=0
> ```

## Update
The command above fixed the `Process failed to launch` error, but
instead produced a `Timeout 30000ms exceeded` error. For now, setting
tests to run on ubuntu-22.04 as that seems to work at the moment.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **Chores**
- Enhanced the automated testing workflow by updating the operating
system to Ubuntu 22.04, improving the reliability of our testing
process.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Baruch Odem <baruchiro@gmail.com>