Build: fix linux build (#641)
Fixes #640: Linux build was failing due to icon. It doesn't seem to work
with `ics` or `ico` formats, or at least not the ones which are used for
Mac/Windows.
To mitigate, set a png icon.

Note that for some reason the icon does not seem to actually apply:

![image](https://github.com/user-attachments/assets/8be0db2d-8459-47fe-a712-e5db72761635)
I thought it might be due to the png dimensions, but I tried setting a
512x512 png instead and the result was the same.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Updated the Linux application build with a fresh icon for enhanced
visual branding.
- **Improvements**
- Expanded the testing matrix to include Windows as an operating system
for running tests.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Baruch Odem <baruchiro@gmail.com>